services/ingest/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.41.0
pydantic==2.6.1
python-dateutil==2.8.2
requests==2.31.0
